+++ quantify_scheduler/backends/mock/__init__.py
@@ -1,0 +1,3 @@
+# Repository: https://gitlab.com/quantify-os/quantify-scheduler
+# Licensed according to the LICENCE file on the main branch
+"""Mock backend for the quantify-scheduler."""