+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Operations - Add `ConditionalReset` gate to Qblox gate library. (!788)
 - Schedule - Fix control flow compilation in certain edge cases, when operation length is zero. (!928)
+- Documentation - Add guides on generic hardware backend and developing a new hardware backend in quantify (!817)
 
 ## 0.18.1 (2024-02-22)